Convert `data` from a subpackage to a module
